# Switching "deep-equal" package for "fast-deep-equal".