--- a/packages/create-sitecore-jss/src/templates/react/package.json
+++ b/packages/create-sitecore-jss/src/templates/react/package.json
@@ -36,7 +36,7 @@
     "axios": "^1.2.0",
     "bootstrap": "^5.2.3",
     "cross-fetch": "^3.1.5",
-    "deep-equal": "^2.1.0",
+    "fast-deep-equal": "3.1.3",
     "graphql": "~16.6.0",
     "graphql-tag": "~2.12.6",
     "i18next": "^22.0.6",