--- a/packages/sitecore-jss-react/package.json
+++ b/packages/sitecore-jss-react/package.json
@@ -64,7 +64,7 @@
   },
   "dependencies": {
     "@sitecore-jss/sitecore-jss": "21.7.0-canary.13",
-    "deep-equal": "^2.1.0",
+    "fast-deep-equal": "3.1.3",
     "prop-types": "^15.8.1",
     "style-attr": "^1.3.0"
   },