<!--- Provide a general summary of your changes in the Title above -->

- [x] This PR follows the [Contribution Guide](https://github.com/Sitecore/jss/blob/dev/CONTRIBUTING.md)
- [ ] Changelog updated

## Description / Motivation
<!--- Describe your changes in detail -->
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->
We were noticing that `deepEqual` in `componentDidUpdate` of `SitecoreContext` was taking up a significant chunk of time for large Layout Service payloads.  

The package "fast-deep-equal" shows 100x faster performance in benchmarks compared to "deep-equal".  

This was the only usage of the "deep-equal" package, so we were able to replace the dependency entirely.

## Testing Details
<!--- Please describe how you tested your changes. -->
<!--- When applicable, include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
- [ ] Unit Test Added
- [x] Manual Test/Other (Please elaborate)

There is no functional change, only performance change, so no new unit tests.  We have implemented this change on 2 different projects and have not seen any issues, just faster performance.

In our Lighthouse tests for one of our projects, we saw the Total Blocking Time reduced from 1090ms to 20ms in Desktop, and from 5170ms to 330ms in mobile.  The results were reproducible, and we saw significant gains in a different project.

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

Not technically a bug, but it's not new functionality either.